multiprocess: fix test_connections even more
There were "random" hangs in test_connection from some workers. The problem was that the "subscriber ping socket" was rep-req, but that the master end (the rep) sent its message from the Messenger ctor and then closed the socket before some workers could receive the rep. This would cause the worker to hang on polling the closed socket. The solution is to replace the rep-req ping socket with push-pull. The only use of the ping socket was to let the master know that all workers are online and receiving messages on the master-workers sub socket. This needs only a push message, not a reply back from master on the worker.