zmq builtin: move zmq_ppoll detection to FindZeroMQ module
There was an issue that when ZeroMQ was found by the module, but then it turned out it wasn't a version with zmq_ppoll, the target libzmq would still already have been created in the find_package module. Because of this, the zeromq builtin then couldn't define the libzmq target anymore. By including the zmq_ppoll detection in the module and making the module fail when it doesn't find it we solve that problem.

It also makes the code in SearchInstalledSoftware.cmake less complicated.