[WIP] try out Neumaier instead of Kahan summation
This checks whether the existing running sum term or the incoming input
added term is smaller and saves the low-order digits of the smallest of
the two. This is equivalent to Kahan summation on a series ordered by
absolute value from large to small.

This gives more consistent results when summing series in different
orders or different partitions.