[RF] [WIP] began refactoring hessian calculator
It's becoming very complex very fast, though, so we're going to shelve this project for now. The problem is that the current function intertwines a lot of things in one loop: gradients, stepsizes, and of course the hessian matrix itself, are all calculated on the go, interleaved with function (likelihood in our case) calls. To be able to make something out of this, this has to be radically decoupled:

- function/likelihood calls separate
- gradients separate
- diagonal matrix elements separate
- off-diagonal elements

In the current code, the off-diagonal elements are relatively decoupled from the rest, but the other stuff is all calculated at the same time for one parameter at a time in one for-loop body. We should come up with a sensible design for decoupling this.