[WIP] add a test that compares RooMinimizerFcn and MinuitFcnGrad
To compile this, one needs the HZy_split workspace by Zef Wolffs; the path is hardcoded in the test case, so that must be adapted to the local path. Also, in the roofit/roofitcore/test/TestStatistics directory, one must clone the RooFitExtensions repo so that it can be compiled into the test (necessary to load the HZy_split workspace).