[RF] add error handling to MultiProcess/TestStatistics based fits
RooFit has a mechanism for flagging function evaluation errors during fitting that is used to steer Minuit out of the problematic part of parameter space. This mechanism was not used in MultiProcess-powered fits. This commit adds that functionality.

For this to work, and especially for it to work as similarly to non-MultiProcess-powered fits as possible, the following changes were also necessary:

- LikelihoodGradientJob must synchronize maxFCN to the workers, because it is used to move out of problematic parts of parameter space. Without this value, the path the fitter takes with and without MultiProcess can be quite different.
- The other part of the problem-area-moving equation is hidden inside RooNaNPacker "magic" nan values that get passed along from functions as if they are nans (i.e. "regular" doubles). We now have support for these special nan values in the MultiProcess-powered classes as well.
- Most importantly, the actual mechanism to log errors, RooAbsReal::logEvalError, is now used where appropriate. Note that this should be considered a bug fix!
- In MinuitFcnGrad, when parameters did not change, we would exit quickly from operator() using a `parameters_changed` flag. However, this circumvented proper error handling, which was done below. This quick exit is hence removed now.
- Another change for consistency with RooNLLVar was that "proper error handling, which was done below" mentioned in the point above. It is now modernized in the same way as in RooMinimizerFcn.
   * Note: it would be much better to factor this out and put it in RooAbsMinimizerFcn for both classes to reuse.

Tests were added to testLikelihoodGradientJob.cxx to trigger the error handling paths.

All affected files were also automatically reformatted.

Logistical note: this is a manual partial cherry-pick of https://github.com/roofit-dev/root/commit/4574832e7c7480b75269b79b50e8cf964f222f96. That commit implemented a bunch of other things as well, some of which were now committed separately, others which are now dropped (mostly debugging code).